fix: SceneItems package name
