+++ src/main/java/io/obswebsocket/community/client/message/request/Request.java
@@ -73,16 +73,16 @@ import io.obswebsocket.community.client.message.request.record.StartRecordReques
 import io.obswebsocket.community.client.message.request.record.StopRecordRequest;
 import io.obswebsocket.community.client.message.request.record.ToggleRecordPauseRequest;
 import io.obswebsocket.community.client.message.request.record.ToggleRecordRequest;
-import io.obswebsocket.community.client.message.request.sceneItems.CreateSceneItemRequest;
-import io.obswebsocket.community.client.message.request.sceneItems.DuplicateSceneItemRequest;
-import io.obswebsocket.community.client.message.request.sceneItems.GetSceneItemColorRequest;
-import io.obswebsocket.community.client.message.request.sceneItems.GetSceneItemEnabledRequest;
-import io.obswebsocket.community.client.message.request.sceneItems.GetSceneItemListRequest;
-import io.obswebsocket.community.client.message.request.sceneItems.GetSceneItemLockedRequest;
-import io.obswebsocket.community.client.message.request.sceneItems.RemoveSceneItemRequest;
-import io.obswebsocket.community.client.message.request.sceneItems.SetSceneItemEnabledRequest;
-import io.obswebsocket.community.client.message.request.sceneItems.SetSceneItemIndexRequest;
-import io.obswebsocket.community.client.message.request.sceneItems.SetSceneItemLockedRequest;
+import io.obswebsocket.community.client.message.request.sceneitems.CreateSceneItemRequest;
+import io.obswebsocket.community.client.message.request.sceneitems.DuplicateSceneItemRequest;
+import io.obswebsocket.community.client.message.request.sceneitems.GetSceneItemColorRequest;
+import io.obswebsocket.community.client.message.request.sceneitems.GetSceneItemEnabledRequest;
+import io.obswebsocket.community.client.message.request.sceneitems.GetSceneItemListRequest;
+import io.obswebsocket.community.client.message.request.sceneitems.GetSceneItemLockedRequest;
+import io.obswebsocket.community.client.message.request.sceneitems.RemoveSceneItemRequest;
+import io.obswebsocket.community.client.message.request.sceneitems.SetSceneItemEnabledRequest;
+import io.obswebsocket.community.client.message.request.sceneitems.SetSceneItemIndexRequest;
+import io.obswebsocket.community.client.message.request.sceneitems.SetSceneItemLockedRequest;
 import io.obswebsocket.community.client.message.request.scenes.CreateSceneRequest;
 import io.obswebsocket.community.client.message.request.scenes.DeleteSceneTransitionOverrideRequest;
 import io.obswebsocket.community.client.message.request.scenes.GetCurrentPreviewSceneRequest;
@@ -187,16 +187,16 @@ import io.obswebsocket.community.client.message.response.record.StartRecordRespo
 import io.obswebsocket.community.client.message.response.record.StopRecordResponse;
 import io.obswebsocket.community.client.message.response.record.ToggleRecordPauseResponse;
 import io.obswebsocket.community.client.message.response.record.ToggleRecordResponse;
-import io.obswebsocket.community.client.message.response.sceneItems.CreateSceneItemResponse;
-import io.obswebsocket.community.client.message.response.sceneItems.DuplicateSceneItemResponse;
-import io.obswebsocket.community.client.message.response.sceneItems.GetSceneItemColorResponse;
-import io.obswebsocket.community.client.message.response.sceneItems.GetSceneItemEnabledResponse;
-import io.obswebsocket.community.client.message.response.sceneItems.GetSceneItemListResponse;
-import io.obswebsocket.community.client.message.response.sceneItems.GetSceneItemLockedResponse;
-import io.obswebsocket.community.client.message.response.sceneItems.RemoveSceneItemResponse;
-import io.obswebsocket.community.client.message.response.sceneItems.SetSceneItemEnabledResponse;
-import io.obswebsocket.community.client.message.response.sceneItems.SetSceneItemIndexResponse;
-import io.obswebsocket.community.client.message.response.sceneItems.SetSceneItemLockedResponse;
+import io.obswebsocket.community.client.message.response.sceneitems.CreateSceneItemResponse;
+import io.obswebsocket.community.client.message.response.sceneitems.DuplicateSceneItemResponse;
+import io.obswebsocket.community.client.message.response.sceneitems.GetSceneItemColorResponse;
+import io.obswebsocket.community.client.message.response.sceneitems.GetSceneItemEnabledResponse;
+import io.obswebsocket.community.client.message.response.sceneitems.GetSceneItemListResponse;
+import io.obswebsocket.community.client.message.response.sceneitems.GetSceneItemLockedResponse;
+import io.obswebsocket.community.client.message.response.sceneitems.RemoveSceneItemResponse;
+import io.obswebsocket.community.client.message.response.sceneitems.SetSceneItemEnabledResponse;
+import io.obswebsocket.community.client.message.response.sceneitems.SetSceneItemIndexResponse;
+import io.obswebsocket.community.client.message.response.sceneitems.SetSceneItemLockedResponse;
 import io.obswebsocket.community.client.message.response.scenes.CreateSceneResponse;
 import io.obswebsocket.community.client.message.response.scenes.DeleteSceneTransitionOverrideResponse;
 import io.obswebsocket.community.client.message.response.scenes.GetCurrentPreviewSceneResponse;
