+++ src/main/java/io/obswebsocket/community/client/message/request/sceneitems/GetSceneItemColorRequest.java
@@ -1,4 +1,4 @@
-package io.obswebsocket.community.client.message.request.sceneItems;
+package io.obswebsocket.community.client.message.request.sceneitems;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,13 +6,13 @@ import lombok.ToString;
 
 @Getter
 @ToString(callSuper = true)
-public class GetSceneItemLockedRequest extends SceneItemRequest {
+public class GetSceneItemColorRequest extends SceneItemRequest {
 
   private final DataWithId requestData;
 
   @Builder
-  private GetSceneItemLockedRequest(String sceneName, Integer sceneItemId) {
-    super(Type.GetSceneItemLocked);
+  private GetSceneItemColorRequest(String sceneName, Integer sceneItemId) {
+    super(Type.GetSceneItemColor);
 
     this.requestData = DataWithId.builder().sceneName(sceneName).sceneItemId(sceneItemId).build();
   }
