+++ src/main/java/io/obswebsocket/community/client/message/request/sceneitems/GetSceneItemLockedRequest.java
@@ -1,4 +1,4 @@
-package io.obswebsocket.community.client.message.request.sceneItems;
+package io.obswebsocket.community.client.message.request.sceneitems;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,13 +6,13 @@ import lombok.ToString;
 
 @Getter
 @ToString(callSuper = true)
-public class RemoveSceneItemRequest extends SceneItemRequest {
+public class GetSceneItemLockedRequest extends SceneItemRequest {
 
   private final DataWithId requestData;
 
   @Builder
-  private RemoveSceneItemRequest(String sceneName, Integer sceneItemId) {
-    super(Type.RemoveSceneItem);
+  private GetSceneItemLockedRequest(String sceneName, Integer sceneItemId) {
+    super(Type.GetSceneItemLocked);
 
     this.requestData = DataWithId.builder().sceneName(sceneName).sceneItemId(sceneItemId).build();
   }
