+++ src/main/java/io/obswebsocket/community/client/message/response/sceneitems/SetSceneItemIndexResponse.java
@@ -1,4 +1,4 @@
-package io.obswebsocket.community.client.message.response.sceneItems;
+package io.obswebsocket.community.client.message.response.sceneitems;
 
 import io.obswebsocket.community.client.message.request.Request;
 import io.obswebsocket.community.client.message.response.RequestResponse;
@@ -7,9 +7,9 @@ import lombok.ToString;
 
 @Getter
 @ToString(callSuper = true)
-public class SetSceneItemEnabledResponse extends RequestResponse {
+public class SetSceneItemIndexResponse extends RequestResponse {
 
-  public SetSceneItemEnabledResponse() {
-    super(Request.Type.SetSceneItemEnabled);
+  public SetSceneItemIndexResponse() {
+    super(Request.Type.SetSceneItemIndex);
   }
 }
